// microservices/emails/src/enums/emails.ts
export enum Emails {
  AFFILIATE_JOINED = "AFFILIATE_JOINED",
  AFFILIATE_REQUISITES_CREATED = "AFFILIATE_REQUISITES_CREATED",
  AFFILIATE_WITHDRAW = "AFFILIATE_WITHDRAW",

  COURSE_BOUGHT = "COURSE_BOUGHT",
  COURSE_FINISHED = "COURSE_FINISHED",
  COURSE_INVITE_SINGLE = "COURSE_INVITE_SINGLE",
  COURSE_PURCHASED = "COURSE_PURCHASED",
  COURSE_SUSPENDED = "COURSE_SUSPENDED",

  HOMEWORK_ANSWER_SENT = "HOMEWORK_ANSWER_SENT",
  HOMEWORK_REVIEWED = "HOMEWORK_REVIEWED",

  MESSAGES_UNREAD = "MESSAGES_UNREAD",

  SCHOOL_GOT_ACCESS = "SCHOOL_GOT_ACCESS",
  SCHOOL_REGISTERED = "SCHOOL_REGISTERED",
  SCHOOL_REQUISITES_APPENDED = "SCHOOL_REQUISITES_APPENDED",
  SCHOOL_WITHDRAW_REQUEST = "SCHOOL_WITHDRAW_REQUEST",

  TARIFF_BEFORE_CHANGE = "TARIFF_BEFORE_CHANGE",
  TARIFF_EXPIRES = "TARIFF_EXPIRES",
  TARIFF_PURCHASED = "TARIFF_PURCHASED",

  USER_EMAIL_CHANGE = "USER_EMAIL_CHANGE",
  USER_EMAIL_CHANGED = "USER_EMAIL_CHANGED",
  USER_PASSWORD_CHANGED = "USER_PASSWORD_CHANGED",
  USER_PASSWORD_RECOVERED = "USER_PASSWORD_RECOVERED",
  USER_REGISTERED = "USER_REGISTERED",

  WEBINAR_COMING = "WEBINAR_COMING",
}